+++ crates/cli/src/commands/unalias.rs
@@ -1,9 +1,10 @@
 use crate::helpers::{map_pin_type, PinOption};
 use crate::session::ProtoSession;
 use clap::Args;
+use iocraft::element;
 use proto_core::{Id, ProtoConfig};
 use starbase::AppResult;
-use starbase_styles::color;
+use starbase_console::ui::*;
 
 #[derive(Args, Clone, Debug)]
 pub struct UnaliasArgs {
@@ -13,14 +14,7 @@ pub struct UnaliasArgs {
     #[arg(required = true, help = "Alias name")]
     alias: String,
 
-    #[arg(
-        long,
-        group = "pin",
-        help = "Remove from the global ~/.proto/.prototools"
-    )]
-    global: bool,
-
-    #[arg(long, group = "pin", help = "Location of .prototools to remove from")]
+    #[arg(long, help = "Location of .prototools to remove from")]
     from: Option<PinOption>,
 }
 
@@ -30,8 +24,7 @@ pub async fn unalias(session: ProtoSession, args: UnaliasArgs) -> AppResult {
     let mut value = None;
 
     let config_path = ProtoConfig::update(
-        tool.proto
-            .get_config_dir(map_pin_type(args.global, args.from)),
+        tool.proto.get_config_dir(map_pin_type(false, args.from)),
         |config| {
             if let Some(tool_configs) = &mut config.tools {
                 if let Some(tool_config) = tool_configs.get_mut(&tool.id) {
@@ -44,21 +37,35 @@ pub async fn unalias(session: ProtoSession, args: UnaliasArgs) -> AppResult {
     )?;
 
     let Some(value) = value else {
-        eprintln!(
-            "Alias {} not found in config {}",
-            color::id(&args.alias),
-            color::path(config_path)
-        );
+        session.console.render(element! {
+            Notice(variant: Variant::Caution) {
+                StyledText(
+                    content: format!(
+                        "Alias <label>{}</label> for <id>{}</id> not found in config <path>{}</path>",
+                        args.alias,
+                        args.id,
+                        config_path.display()
+                    ),
+                )
+            }
+        })?;
 
         return Ok(Some(1));
     };
 
-    println!(
-        "Removed alias {} ({}) from config {}",
-        color::id(&args.alias),
-        color::muted_light(value.to_string()),
-        color::path(config_path)
-    );
+    session.console.render(element! {
+        Notice(variant: Variant::Success) {
+            StyledText(
+                content: format!(
+                    "Removed <id>{}</id> alias <label>{}</label> <mutedlight>({})</mutedlight> from config <path>{}</path>",
+                    args.id,
+                    args.alias,
+                    value.to_string(),
+                    config_path.display()
+                ),
+            )
+        }
+    })?;
 
     Ok(None)
 }